Added icons
Through the react-native-cross-platform-icons component, the necessary icons were added to simplify texts in the app.

A react-native wrapper component that displays the appropriate icon based on the current platform ( Android and iOS )

Installation

    npm install --save react-native-cross-platform-icons react-native-vector-icons

After Installation (this step is required)

    react-native link react-native-vector-icons

Usage

    import CrossPlatformIcon from 'react-native-cross-platform-icons';

    <CrossPlatformIcon name="home" size={45} color="black" outline />
    <CrossPlatformIcon name="arrow-back" size={40} color="#BADA55" />

The indicative name, allows us to indicate the icon to be displayed, based on the page:

    https://ionicons.com/

The size and color refer to that of the typographic font.

The icons for the App were also added, from Android Studio.

ToDo:

— The position of the favorite icon must be corrected.
— The onClick function must be implemented to the favorite icon.