Favorite button position correction
Fixed the position of the Favorite button so that it is positioned permanently at the bottom right.

The React Native function, resolveAssetSource, was also added to set it as the main one when calling the house image.

Finally, a README.md was added to the project, with the specifications to be able to execute it.